Drop (uint16_t)e->count truncation in slot bounds checks
The pattern `slot < (uint16_t)e->count` truncated Env.count (int) to
16 bits before comparison. For env counts ≤ 65535 the cast is a no-op;
above that, valid slots in the truncated range would silently read as
NULL. Unreachable in practice — scripts don't have >65535 globals or
function locals — but the cast adds nothing and obscures intent.
Compare `(int)slot < e->count` directly across all 7 sites
(GET_LOCAL, SET_LOCAL, LOCAL_DOT_GET/SET, LOCAL_IDX_GET,
LOCAL_IDX_DOT_GET/SET).

Suite: 1030/1030 PASS. DMG cpu_instrs: 0.41 MHz, serial "cpu_instrs".

+++ src/vm.c
@@ -560,7 +560,7 @@ static Value *vm_run(EigsChunk *chunk, Env *env) {
         uint16_t slot = read_u16(ip); ip += 2;
         /* Read from the function's original env (not loop-fresh child) */
         Env *e = frame->fn_env;
-        if (slot < (uint16_t)e->count) {
+        if ((int)slot < e->count) {
             Value *v = e->values[slot];
             if (v) { val_incref(v); vm_push(v); }
             else vm_push(make_null());
@@ -575,7 +575,7 @@ static Value *vm_run(EigsChunk *chunk, Env *env) {
         Value *v = vm_peek(0);
         /* Write to the function's original env (not loop-fresh child) */
         Env *e = frame->fn_env;
-        if (slot < (uint16_t)e->count) {
+        if ((int)slot < e->count) {
             val_incref(v);
             val_decref(e->values[slot]);
             e->values[slot] = v;
@@ -1022,7 +1022,7 @@ static Value *vm_run(EigsChunk *chunk, Env *env) {
         uint16_t slot = read_u16(ip); ip += 2;
         uint16_t name_idx = read_u16(ip); ip += 2;
         Env *e = frame->fn_env;
-        Value *target = (slot < (uint16_t)e->count) ? e->values[slot] : NULL;
+        Value *target = ((int)slot < e->count) ? e->values[slot] : NULL;
         Value *result = make_null();
         if (target && target->type == VAL_DICT) {
             const char *key = chunk->constants[name_idx]->data.str;
@@ -1045,7 +1045,7 @@ static Value *vm_run(EigsChunk *chunk, Env *env) {
         uint16_t name_idx = read_u16(ip); ip += 2;
         Value *val = vm_peek(0);
         Env *e = frame->fn_env;
-        Value *target = (slot < (uint16_t)e->count) ? e->values[slot] : NULL;
+        Value *target = ((int)slot < e->count) ? e->values[slot] : NULL;
         if (target && target->type == VAL_DICT) {
             const char *key = chunk->constants[name_idx]->data.str;
             uint32_t h = chunk->const_hashes ? chunk->const_hashes[name_idx] : 0;
@@ -1066,7 +1066,7 @@ static Value *vm_run(EigsChunk *chunk, Env *env) {
         uint16_t slot = read_u16(ip); ip += 2;
         uint16_t idx = read_u16(ip); ip += 2;
         Env *e = frame->fn_env;
-        Value *target = (slot < (uint16_t)e->count) ? e->values[slot] : NULL;
+        Value *target = ((int)slot < e->count) ? e->values[slot] : NULL;
         Value *result = make_null();
         if (target) {
             int i = (int)idx;
@@ -1109,7 +1109,7 @@ static Value *vm_run(EigsChunk *chunk, Env *env) {
         uint16_t list_idx = read_u16(ip); ip += 2;
         uint16_t name_idx = read_u16(ip); ip += 2;
         Env *e = frame->fn_env;
-        Value *target = (slot < (uint16_t)e->count) ? e->values[slot] : NULL;
+        Value *target = ((int)slot < e->count) ? e->values[slot] : NULL;
         Value *result = make_null();
         int i = (int)list_idx;
         if (target && target->type == VAL_LIST) {
@@ -1145,7 +1145,7 @@ static Value *vm_run(EigsChunk *chunk, Env *env) {
         uint16_t name_idx = read_u16(ip); ip += 2;
         Value *val = vm_peek(0);
         Env *e = frame->fn_env;
-        Value *target = (slot < (uint16_t)e->count) ? e->values[slot] : NULL;
+        Value *target = ((int)slot < e->count) ? e->values[slot] : NULL;
         int i = (int)list_idx;
         if (target && target->type == VAL_LIST) {
             if (i < target->data.list.count) {
